Adds function caching, cleans up code
